% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepCens-package.R
\docType{package}
\name{DepCens-package}
\alias{DepCens-package}
\alias{DepCens}
\title{The 'DepCens' package.}
\description{
Dependent censoring regression models for survival multivariate data. These models are based on extensions of the frailty models, capable to accommodating the dependence between failure and censoring times, with Weibull and piecewise exponential marginal distributions. Theoretical details regarding the models implemented in the package can be found in Schneider et al. (2019) <doi: 10.1002/bimj.201800391>.
}
