% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{follic}
\alias{follic}
\title{Follicular Cell Lymphoma}
\format{
A data frame with 541 rows and 7 variables:
\itemize{
  \item age:	age
  \item clinstg: clinical stage: 1=stage I, 2=stage II
  \item ch:	chemotherapy
  \item rt:	radiotherapy
  \item hgb: hemoglobin level in g/l
  \item time:	first failure time
  \item status:	censoring status; 0=censored, 1=relapse, 2=death
}
}
\description{
Competing risk data set involving follicular cell lymphoma. The data consist of 541 patients with early disease stage I or II.
}
