% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblio_function_EMMC_Weibull.R
\name{model_MEP_dep}
\alias{model_MEP_dep}
\alias{Piecewise}
\alias{exponential}
\alias{model}
\alias{for}
\alias{dependent}
\alias{censoring}
\title{model_MEP_dep}
\usage{
model_MEP_dep(formula, data, delta_t, delta_c, ident, Num_intervals)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): should be used as 'time ~ failure covariates | informative covariates'.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model.}

\item{delta_t}{Indicator function of the event of interest.}

\item{delta_c}{Indicator function of the dependent censoring.}

\item{ident}{Cluster indicator variable.}

\item{Num_intervals}{Number of intervals of the time grid.}
}
\value{
model_MEP_dep returns an object of class "dcensoring" containing the fitted model.
}
\description{
model_MEP_dep function estimates the parameters of the Piecewise exponential model with dependent censoring, considering the frailty model to capture the dependence between failure and dependent censoring times, as well as the clusters variability
}
\examples{
\dontrun{
# MEP approach:
model_MEP_dep(formula = time ~ x1 | x3, data=KidneyMimic, delta_t=KidneyMimic$delta_t,
              delta_c=KidneyMimic$delta_c, ident=KidneyMimic$ident, Num_intervals=15)
}
}
