% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblio_function_EMMC_Weibull.R
\name{model_Weibull_dep}
\alias{model_Weibull_dep}
\title{model_Weibull_dep}
\usage{
model_Weibull_dep(formula, data, delta_t, delta_c, ident)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): should be used as 'time ~ failure covariates | informative covariates'.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model.}

\item{delta_t}{Indicator function of the event of interest.}

\item{delta_c}{Indicator function of the dependent censoring.}

\item{ident}{Cluster indicator variable.}
}
\value{
model_Weibull_dep returns an object of class "dcensoring" containing the fitted model.
}
\description{
model_Weibull_dep function estimates the parameters of the Weibull model with dependent censoring, considering the frailty model to capture the dependence between failure and dependent censoring times, as well as the clusters variability
}
\examples{
\dontrun{
model_Weibull_dep(formula = time ~ x1 | x3, data=KidneyMimic, delta_t=KidneyMimic$delta_t,
                  delta_c=KidneyMimic$delta_c, ident=KidneyMimic$ident)
}
}
