% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependent_censoring.R
\name{dependent.censoring}
\alias{dependent.censoring}
\title{Dependent Censoring model}
\usage{
dependent.censoring(
  formula,
  data,
  delta_t,
  delta_c,
  ident,
  approach = c("weibull", "mep"),
  Num_intervals = 10
)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): should be used as 'time ~ failure covariates | informative covariates'.}

\item{data}{a data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model.}

\item{delta_t}{Indicator function of the event of interest.}

\item{delta_c}{Indicator function of the dependent censoring.}

\item{ident}{Cluster indicator variable.}

\item{approach}{approach to be used to fit the model (must be either weibull or mep).}

\item{Num_intervals}{Number of intervals of the time grid (mep only).}
}
\value{
dependent.censoring returns an object of class "dcensoring" containing the fitted model.
}
\description{
dependent.censoring is used to fit a survival analysis with dependent censoring.
}
\details{
function estimates the parameters of the Piecewise exponential model (approach = "mep") or Weibull model (approach="weibull") with dependent censoring, considering the frailty model to capture the dependence between failure and dependent censoring times, as well as the clusters variability.
}
\examples{
\dontrun{
library(DepCens)
delta_t <- ifelse(KidneyMimic$cens==1,1,0)
delta_c <- ifelse(KidneyMimic$cens==2,1,0)
fit <- dependent.censoring(formula = time ~ x1 | x3, data=KidneyMimic, delta_t=delta_t,
                          delta_c=delta_c, ident=KidneyMimic$ident, approach = "mep")
summary_dc(fit)
}
}
